Change name of Logout file and component to SignOut
